requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
openai>=1.40.0
llama-index>=0.11.14
llama-index-llms-azure-openai>=0.2.0
llama-index-embeddings-huggingface>=0.3.1
llama-index-readers-file>=0.2.2
faiss-cpu>=1.8.0.post1
sentence-transformers>=3.0.1
pydantic>=2.8.2
